refactored code to use classes
