+++ .gitignore
@@ -1,1 +1,1 @@
-lib/application.yml
+app/lib/application.yml
