+++ app/bin/msgPro
@@ -1,3 +1,3 @@
 #!/usr/bin/env ruby
-require_relative '../app/config/environment'
+require_relative '../config/environment'
 Cli.new.run(ARGV[0], ARGV[1])
