+++ app/data_fetchers/api_caller.rb
@@ -1,0 +1,7 @@
+class ApiCaller
+  attr_reader :data
+  def initialize(zip)
+    url = "http://api.openweathermap.org/data/2.5/weather?units=imperial&zip=#{zip},us"
+    @data = JSON.parse(open(url).read)
+  end
+end
